pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fferyman"
version = "0.1.0"
description = "Pluggable folder-mirror framework with SQLite-backed mappings"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
fferyman = "fferyman.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
